rust: only bridgestan_download_src is to be feature gated
