+++ rust/Cargo.toml
@@ -15,11 +15,11 @@ ureq = { version = "2.7", optional = true }
 tar = { version = "0.4", optional = true }
 flate2 = { version = "1.0", optional = true }
 dirs = { version = "5.0", optional = true }
-path-absolutize = { version = "3.1", optional = true }
-log = { version = "0.4", optional = true }
+path-absolutize = { version = "3.1" }
+log = { version = "0.4" }
 
 [features]
-compile-stan-model = ["dep:ureq", "dep:tar", "dep:flate2", "dep:dirs", "dep:path-absolutize", "dep:log"]
+download-bridgestan-src = ["dep:ureq", "dep:tar", "dep:flate2", "dep:dirs"]
 
 [build-dependencies]
 bindgen = "0.69.1"
@@ -28,7 +28,8 @@ bindgen = "0.69.1"
 approx = "0.5.1"
 rand = "0.8.5"
 env_logger = "0.11"
+dirs = { version = "5.0" }
 
 [[example]]
 name = "example"
-required-features = ["compile-stan-model"]
+required-features = ["download-bridgestan-src"]
